+++ BCInterpreter1/src/FrameStack.hpp
@@ -16,21 +16,21 @@ class FrameStack
         void SetGlobalFrame(int i);
         void PushStack(int i);
         void PopStack();
-        void SetLocal(uint32_t index, uint32_t pointer);
+        void SetLocal(uint16_t index, uint64_t pointer);
         void SetReturn(uint32_t ret);
         uint32_t GetReturn();
-        uint32_t GetLocal(uint32_t index);
-        void SetGlobal(const string & name, uint32_t pointer);
-        uint32_t GetGlobal(const string & name);
-        void InsertGlobal(const string & name, uint32_t index);
+        uint64_t GetLocal(uint16_t index);
+        void SetGlobal(const string & name, uint64_t pointer);
+        uint64_t GetGlobal(const string & name);
+        void InsertGlobal(const string & name, uint64_t index);
     private:
         struct Frame{
-            vector<uint16_t> m_locals;
+            vector<uint64_t> m_locals;
             uint32_t m_ret;
             Frame * m_prev = nullptr;
         };
         struct GlobalFrame{
-            unordered_map<string,uint32_t> m_GlobalMap;
+            unordered_map<string,uint64_t> m_GlobalMap;
         };        
         GlobalFrame m_global;
         Frame * m_local = nullptr;
